requirements.txt
dash>=2.18.0
dash-bootstrap-components>=1.6.0
databricks-sql-connector>=3.0.0
databricks-sdk>=0.28.0
plotly>=5.18.0
pandas>=2.0.0
python-dotenv>=1.0.0
